Change decimals to doubles

+++ Assets/Scripts/DamageReport.cs
@@ -2,5 +2,5 @@ public class DamageReport
 {
     public EntityController causedBy;
     public EntityController target;
-    public decimal damage;
+    public double damage;
 }
