+++ Assets/Scripts/Stats/Counter.cs
@@ -23,7 +23,7 @@ public class Counter
         currentValue = defaultValue;
     }
 
-    public void Reset(decimal newDefault) => Reset((float)newDefault);
+    public void Reset(double newDefault) => Reset((float)newDefault);
 
     public void Reset(Stat newDefault)
     {
